+++ includes/requirements.php
@@ -58,7 +58,7 @@ function wps_requirements_check( $force_check = false ) {
 
 
 function wps_notice_aihrus() {
-	$help_url  = esc_url( 'https://aihrus.zendesk.com/entries/35689458' );
+	$help_url  = esc_url( 'https://nodedesk.zendesk.com/hc/en-us/articles/202381391' );
 	$help_link = sprintf( __( '<a href="%1$s">Update plugins</a>. <a href="%2$s">More information</a>.' ), self_admin_url( 'update-core.php' ), $help_url );
 
 	$text = sprintf( esc_html__( 'Plugin "%1$s" has been deactivated as it requires a current Aihrus Framework. Once corrected, "%1$s" can be activated. %2$s' ), WPS_NAME, $help_link );